test(results): re-anchor the five vacuous hero-rerun pins onto a real sweep
All five `queryByTestId('hero-rerun')` negative pins asserted the absence of
a testid that exists NOWHERE in source — they asserted the absence of
something that could never be present, so they could never fail. A pin that
cannot fail is not a pin.

Re-anchored onto `collectRerunControls`, which sweeps a subtree for any
control matching a rerun-ish accessible NAME or TESTID, so a reintroduction
has to evade both. Where a hero root exists the pins now anchor to
`getByTestId('analysis-hero-panel')`, which throws if the hero stops
rendering — the pin cannot silently go vacuous again.

The vocabulary lives in one place (tests/helpers/rerunControls.ts) rather
than five drifting copies, which is what let the dock spec match only the
literal /rerun|re-run/ while the hero specs matched a dead testid.

Verified the re-anchored pins BITE: injecting a plain `<button>Run again` into
AnalysisHeroPanel fails all four hero pins — a control the old testid pins
would have missed completely, since it carries no testid at all. Reverted the
injection; suites green (277 tests, incl. the analysis-hero inertness guard).
