+++ src/components/results/analysis-hero/__tests__/a11y.spec.tsx
@@ -10,6 +10,7 @@ import { AnalysisHeroPanel } from '../AnalysisHeroPanel'
 import { buildHeroModel } from '../buildHeroModel'
 import type { HeroChartModel, HeroStatusModel } from '../heroTypes'
 import { FULL_COMPLETENESS, makeHeroData } from '../__fixtures__/hero.fixtures'
+import { collectRerunControls } from '../../../../../tests/helpers/rerunControls'
 
 function chartModel(): HeroChartModel {
   return buildHeroModel(makeHeroData()) as HeroChartModel
@@ -122,7 +123,11 @@ describe('AnalysisHero — accessibility', () => {
     // RETIRED PIN: the hero's stale Re-run pill is removed — the freshness
     // strip owns the one Rerun; the panel is staleness-agnostic.
     const { container } = renderPanel(chartModel())
-    expect(screen.queryByTestId('hero-rerun')).toBeNull()
+    // Re-anchored (C1 review): this asserted `queryByTestId('hero-rerun')`
+    // was absent, but that testid exists nowhere in source — it asserted the
+    // absence of something that could never be present, so it could never
+    // fail. Sweep the rendered hero for a run control of ANY name/testid.
+    expect(collectRerunControls(container)).toEqual(new Set())
     // No aria-disabled/inert wrapper: rows and tabs remain in the a11y tree.
     expect(screen.getByTestId('hero-lens-tab-goal')).toBeEnabled()
     await expectNoViolations(container)
